[processor] Capture turbostat output
The turbostat command is a topological report. This report will show
what cstate and pstate processors are in, what the processor frequency
is during that timeand how busy each core is. The cstate %'s tell us
when we are saving power at the expense of performance which is useful
for knowing if the host is optimally configured.

Fixes: #802

Signed-off-by: Shane Bradley <sbradley@redhat.com>